Commented Gravity and ForceField. Finished writing Coord, Vec2D, Gravity, and ForceField
I finished writing the code for classes Coord, Vec2D, Gravity, and
ForceField. This code will most likely not be touched again until I
finish the framework and start testing/debugging it. I also wrote
comments for all of these classes detailing what the classes are for,
what their instance variables are for, and what their methods do/how
they work. I also made some minor edits to the Surface class. I forgot
that there are two different kinds of friction, kinetic and static, and
that the Surface class needs variables for the coefficients of both
kinds of friction. The methods hit() and slide() for the Surface class
also needed another parameter for the collision ID of the tile that
collided with the RBObject calling the hit() and/or slide() method of
the surface since a single Surface object can be used for multiple
kinds of tiles. I've finished all the purely physics related classes in
my framework except for the Surface class. The Surface class seems like
it will be the hardest class to code in the entire framework, so I'm
saving it to work on over the upcoming long weekend.