Edits to Vec2D and RBObject.
I decided that a method that returns an exact copy of a Vec2D object
might be useful for making a new Vec2D object with the same magnitude
and direction as another. I also decided that since velocity,
acceleration, and jerk are properties that are constantly affected by
other objects around an RBObject, it makes more sense to make them
public.