Finished InputManager (but had to move code to GameFrame)
I was working on the InputManager class that tracks user input from the
keyboard and mouse when I came across a point where I had to use a
method from JGEngine (which GameFrame extends) to actually see the
current state of a button. I thought the method would be static and
that I could access it by typing 'GameFrame.getKey(key)', but it turned
out that the method wasn't static and I could not use it without first
having an instance of GameFrame (which is impossible because GameFrame
is abstract). I ended up having to move the code for InputManager into
the GameFrame class to get it to work. So now the GameFrame class keeps
track of its own input instead of having another class do that for it.
After moving the code for the InputManager class into the GameFrame
class, I finished and fully commented it. I will finish the rest of the
GameFrame class later.