Completed and commented ManagerList and ActorManager
I worked on writing the code for ManagerList and ActorManager today,
and commented all their methods and instance variables along with
descriptions of how the classes are used. ManagerList came out working
how I thought it would. When I started writing code for ActorManager
though, I quickly realized that I could not actually write code for how
the class should spawn objects without knowing exactly what kinds of
Actor objects the class was going to create (which is impossible for it
know because it will vary from game to game that uses GameFrame as a
framework). I decided to make ActorManager abstract and require that
someone using the class extend it and then fill in their own code for
the spawn() method. Some other things I changed were that I made some
instance variables for ActorManager public so that ManagerList could
change them and I made currentFrame in GameFrame along with
getCurrFrame() static so that other classes could access getCurrFrame()
without having an instance of GameFrame. This is okay since there will
only ever be one instance of GameFrame anyway using the variable
currentFrame.