Finished RBObject class, updated constructors.
I finished writing and commenting all the code for the RBObject class.
It is now capable of colliding with other RBObjects as well as tiles
and having physical interactions with them. It also reacts to force
fields as well. I noticed as I was working on the RBObject class though
that the constructors did not set the bypassTile setting, so I updated
all the constructors for RBObject, Particle, Projectile, and Pickup to
address this setting.