Worked on Actor class, edited RBObject
I worked on the Actor class, filling in code for methods and writing
constructors. The Actor class still needs to be commented, and the code
for the subclass Extension still needs to be written and commented. I
also edited the RBObject class so that the GameFrame class can call the
move() method twice, once to implement the effects of gravity and force
fields and once to update the RBObjects velocity. The reason I did this
is because all of those tasks are tasks that GameFrame has to carry out
for each RBObject every frame, but some of them before collision
detection and some after. The way the JGame engine works though (which
GameFrame is built around), for such code to be run every frame for
each object, it must go into the move() method. JGame does not support
being able to call other methods of a JGObject every frame aside from
move() and any collision methods that are triggered by a collision
during the frame being executed. So I was forced to recycle the move()
method and have it called twice, even though it would have been more
convenient to have two separate methods called each frame, one before
collision detection and one after.