Finished the GameFrame class
Finished coding and commenting the GameFrame class. The comments
explain a bit about what the GameFrame framework is for and how to use
it, although the comments refer you to the commenting in other classes
to know how they work. Written in the comments is also an established
convention when using GameFrame, and that is to call the super method
at the very end of any method that you override (and not the
beginning). Now only have one class left to write (the Surface class).