Finished Surface class. Completed framework.
I finished coding and commenting the Surface class, which was the last
class in my framework which I had to write. I also had to make some
edits to the RBObject class to make it work with the Surface class.
Next step is to create a tech demo of the framework and debug as many
problems with the framework as I can.