Some edits to Surface, Actor, and Projectile classes, started tech demo.
I added a new method to the Actor and Projectile classes that someone
using the GameFrame framework can call in order to have one of those
objects accelerate themselves while taking into account that the object
has a maximum speed it may induce on itself.

I also realized I forgot to make the Surface class work when the
Surface is considered to be in the background, so I started fixing that
as well.

Finally, I started working on a tech demo that shows how to use the
GameFrame framework. I also imported into my project sprites that I
threw together from scratch the other night that I will be using in my
tech demo game and created a media table file for those sprites and
their animations.