Some more changes to Surface class. Set up play field for tech demo.
I started putting together what the layout of the tiles (environment)
would be in the play field for the tech demo game. I finished arranging
the tiles in what looked like an interesting arrangement for the game.
I also made it so that a crosshairs image follows your mouse cursor,
but I was not able to get the original mouse cursor to disappear. I
also created some new classes, Launcher, Ball, Steel, and Railing that
will extend classes from the GameFrame framework to create game objects
and surfaces for the tiles in the tech demo. While working on the tech
demo, I realized that the origin in JGame is in the upper left corner
of the play field, so I had to go back and change some code in the
Surface class in order to make it work with an inverted y-axis. I also
had to change the standard for measuring angle directions by saying
that GameFrame uses and inverted unit circle to measure angles (0
radians still points right, but pi/2 points down instead of up).