Fixed Collision Bugs! Continuing on making tech demo.
I fixed the last couple of bugs in my code for calculating collisions.
Turns out whenever I was creating a new vector objects to return as an
impulse, I used the wrong constructor (for a class I created, how
embarrassing). Next, I'm going to code the Surface objects in my tech
demo to create particles when they come in contact with things.