Started working on fixing the last collision bug, but wanted to make a commit before doing so.
There is one last collision bug. RBObjects sometimes pass through walls
because the RBObject thinks it has hit the the top of a lower tile when
it should have decided it was hitting the left or right side of the
tile above the lower one in the wall. A similar thing happens when
hitting floors. I know how to fix this, but it will take major
alterations to my code for the Surface class. I also tried to
incorporate a new game object into the tech demo that creates an
electric force field around itself, but whenever the objects spawn they
seem to get removed from the game before they are even able to appear
in the next frame. I have no ideas as of now why this is or how to fix
it. It's also a serious bug that jeopardizes the entire framework.
There's no point in having a framework that manages the creation of
objects if those objects are all immediately deleted.